// src/app/shared/model/country.ts
export class Country {
  flag: string;
  name: string;
  alpha2Code: string;
  alpha3Code: string;
  altSpellings: [];
  borders: [{}];
  capital: string;
  currencies: [{}];
  nativeName: string;
  population: number;
  region: string;
  regionalBlocs: [{
    acronym: string;
    name: string;
  }];
  subregion: string;
  topLevelDomain: [string];
  countryN: [{}];
  languages: [{
    name: string;
  }];
}